Реализация методов для работы с таблицами Продукты
