+++ src/main/java/ProductMain.java
@@ -1,15 +1,10 @@
-import dao.ProductDAO;
-import dao.ProductDAOImpl;
-import menu.ProductMenu;
-
-import java.util.Scanner;
+import scanner.ProductServiceMenu;
 
 public class ProductMain {
 
     public static void main (String [] args){
-        ProductDAO productDAO = new ProductDAOImpl();
-        productDAO.createTable();
-        ProductMenu.menu();
+
+        ProductServiceMenu.menu();
 
     }
 }
