+++ src/main/java/entity/Product.java
@@ -1,0 +1,33 @@
+package entity;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+public class Product {
+    //идентификатор
+    private long id;
+    //наименование
+    private String name;
+    //цена
+    private BigDecimal price;
+    //срок годности
+    private LocalDate expiryDate;
+
+    @Override
+    public String toString() {
+        return "Product{" +
+                "id=" + id +
+                ", name='" + name + '\'' +
+                ", price=" + price +
+                ", expiryDate=" + expiryDate +
+                '}';
+    }
+
+}
