+++ src/main/java/scanner/ProductServiceMenu.java
@@ -1,55 +1,72 @@
-package menu;
+package scanner;
 
-import entities.Product;
-import service.ProductService;
+import dao.ProductDAO;
+import dao.impl.ProductDAOImpl;
+import entity.Product;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Scanner;
 
-public class ProductMenu {
+public class ProductServiceMenu {
     private static final Scanner scanner = new Scanner(System.in);
-    private static ProductService productService = new ProductService();
+    private static ProductDAO productRepository = new ProductDAOImpl();
 
     public static void menu() {
         while (true) {
             System.out.println("Выберите действие для Товара:");
-            System.out.println("1. Создать продукт");
-            System.out.println("2. Найти продукт по ID");
-            System.out.println("3. Найти продукт по наименованию");
-            System.out.println("4. Вывести список всех продуктов");
-            System.out.println("5. Обновить информацию о продукте");
-            System.out.println("6. Удалить продукт");
-            System.out.println("7. Вернуться в главное меню");
+            System.out.println("1. Создать таблицу продуктов");
+            System.out.println("2. Удалить таблицу продуктов");
+            System.out.println("3. Создать продукт");
+            System.out.println("4. Найти продукт по id");
+            System.out.println("5. Найти продукт по наименованию");
+            System.out.println("6. Вывести список всех продуктов");
+            System.out.println("7. Обновить информацию о продукте");
+            System.out.println("8. Удалить продукт");
+            System.out.println("9. Вернуться в главное меню");
             int actionChoice = scanner.nextInt();
             scanner.nextLine();
 
             switch (actionChoice) {
                 case 1:
-                    createProduct();
+                    createTable();
                     break;
                 case 2:
-                    findProductById();
+                    dropTable();
                     break;
                 case 3:
-                    findProductByName();
+                    createProduct();
                     break;
                 case 4:
-                    findAllProducts();
+                    findProductById();
                     break;
                 case 5:
-                    updateProduct();
+                    findProductByName();
                     break;
                 case 6:
-                    deleteProduct();
+                    findAllProducts();
                     break;
                 case 7:
+                    updateProduct();
+                    break;
+                case 8:
+                    deleteProduct();
+                    break;
+                case 9:
                     return; // Возврат в главное меню
                 default:
                     System.out.println("Неверный выбор. Попробуйте снова.");
             }
         }
+    }
+
+    public static void createTable(){
+        productRepository.createTable();
+    }
+
+    public static void dropTable(){
+        productRepository.dropTable();
     }
 
     public static void createProduct() {
@@ -58,29 +75,14 @@ public class ProductMenu {
         System.out.println("Введите имя товара:");
         String name = scanner.nextLine();
 
-        BigDecimal price;
-        while (true) {
-            System.out.println("Введите цену товара");
-            if (scanner.hasNextBigDecimal()) {
-                price = scanner.nextBigDecimal();
-                break;
-            } else {
-                System.out.println("Некорректный ввод ");
-                scanner.next();
-            }
-        }
-        scanner.nextLine();
+        System.out.println("Введите цену товара");
+        BigDecimal price = scanner.nextBigDecimal();
 
-        LocalDate expireDate;
-        while (true) {
-            System.out.println("Введите срок годности товара");
-            expireDate = LocalDate.parse(scanner.next());
-            break;
-        }
         scanner.nextLine();
-
-        productService.create(id, name, price, expireDate);
-        System.out.println("Продукт с id: " + id + " создан");
+        System.out.println("Введите срок годности товара");
+        LocalDate expireDate =LocalDate.parse(scanner.nextLine());
+        //scanner.nextLine();
+        productRepository.create(new Product (id, name, price, expireDate));
     }
 
     private static void findProductById() {
@@ -88,7 +90,7 @@ public class ProductMenu {
         long id = scanner.nextLong();
         scanner.nextLine();
 
-        Product product = productService.getById(id);
+        Product product = productRepository.getById(id);
         if (product != null) {
             System.out.println("Найден продукт: " + product);
         } else {
@@ -102,7 +104,7 @@ public class ProductMenu {
         String name = scanner.next();
         scanner.nextLine();
 
-        List<Product> products = productService.getByName(name);
+        List<Product> products = productRepository.getByName(name);
         if (products != null) {
             System.out.println("Найдены продукт: ");//TODO
             //распечатать список продуктов
@@ -114,7 +116,7 @@ public class ProductMenu {
 
     private static void findAllProducts() {
         System.out.println("Список всех продуктов:");
-        productService.findAllProducts().forEach(System.out::println);
+        productRepository.getAllProducts().forEach(System.out::println);
     }
 
     private static void updateProduct() {
@@ -122,7 +124,7 @@ public class ProductMenu {
         long id = scanner.nextLong();
         scanner.nextLine();
 
-        Product product = productService.getById(id);
+        Product product = productRepository.getById(id);
         if (product == null) {
             System.out.println("Продукт с id " + id + " не найден");
             return;
@@ -138,7 +140,7 @@ public class ProductMenu {
         System.out.println("Введите новый срок годности");
         LocalDate newExpireDate = LocalDate.parse(scanner.next());
 
-        productService.update(new Product(id, newName, newPrice, newExpireDate));
+        productRepository.update(new Product(id, newName, newPrice, newExpireDate));
         System.out.println("Товар обновлен: " + product);
     }
 
@@ -147,7 +149,7 @@ public class ProductMenu {
         long id = scanner.nextLong();
         scanner.nextLine();
 
-        productService.delete(id);
+        productRepository.delete(id);
         System.out.println("Покупатель с ID " + id + " удален");
     }
